requirements.txt
beautifulsoup4
bs4
Django
django-cors-headers
django-filter
django-froala-editor
django-jazzmin
django-js-asset
django-storages @ git+https://github.com/rohanphulkar/django-storages
djangorestframework
djangorestframework-simplejwt
gunicorn
mysqlclient
Pillow
pycryptodome
PyJWT
pypdf
pypng
python-decouple
pytz
PyYAML
qrcode
requests
shortuuid
six
sqlparse
urllib3==2.0.5
uuid
xhtml2pdf
